+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - BusinessIdentity and BusinessAttachment resources
 - PixPullSubscription resource
 - PixPullRequest resource
+- IssuingStockRule resource
 
 ## [0.16.0] - 2026-06-01
 ### Added